[common] Move Linear and ScaledBiasRowwiseLinear into common/linear.py
Introduce torchtitan/models/common/linear.py as the home for configurable
linear modules. Move the Linear wrapper out of nn_modules.py and
ScaledBiasRowwiseLinear out of gpt_oss/model.py into this new module, and
repoint all import sites accordingly.

- nn_modules.py no longer defines Linear (removed from __all__).
- gpt_oss/model.py drops the now-unused torch.nn.functional import.
- common/__init__.py re-exports Linear and ScaledBiasRowwiseLinear.
- gpt_oss/__init__.py imports ScaledBiasRowwiseLinear from common.linear.

Pure move with no behavior change.

Test Plan:
pre-commit run on all changed files: passed (flake8, ufmt, pyrefly, etc.).
Ran affected unit tests, all green: test_linear, test_module, test_train_spec,
test_lora, test_fused_qkv, test_rope, test_quantization,
test_activation_checkpoint, test_fused_swiglu, test_compile_moe,
test_compile_regional_inductor.
