+++ tests/unit_tests/test_quantization.py
@@ -10,8 +10,8 @@ from torchtitan.components.quantization.float8 import _get_float8_grouped_expert
 from torchtitan.components.quantization.mx import _get_mxfp8_grouped_experts_cls
 from torchtitan.components.quantization.utils import has_quantization
 from torchtitan.config import ConfigManager
+from torchtitan.models.common.linear import Linear
 from torchtitan.models.common.moe import GroupedExperts
-from torchtitan.models.common.nn_modules import Linear
 from torchtitan.models.gpt_oss.moe import GptOssGroupedExperts
