+++ torchtitan/models/flux/model/layers.py
@@ -13,7 +13,8 @@ import torch
 from einops import rearrange
 from torch import nn, Tensor
 from torchtitan.models.common.attention import ScaledDotProductAttention
-from torchtitan.models.common.nn_modules import GELU, LayerNorm, Linear, RMSNorm, SiLU
+from torchtitan.models.common.linear import Linear
+from torchtitan.models.common.nn_modules import GELU, LayerNorm, RMSNorm, SiLU
 from torchtitan.protocols.module import Module, Sequential
